refactor: 도메인과 뷰, 컨트롤러를 분리
