+++ src/main/kotlin/racingcar/Main.kt
@@ -1,13 +1,7 @@
 package racingcar
 
-import racingcar.domain.RacingGameParam
-import racingcar.io.InputView
-import racingcar.racing.RacingGame
+import racingcar.controller.RacingGameController
 
 fun main() {
-    val carNameList = InputView.getCarNames()
-    val rounds = InputView.getRoundCount()
-
-    val gameParam = RacingGameParam(carNameList, rounds)
-    RacingGame(gameParam).start()
+    RacingGameController.start()
 }
