+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -17,14 +17,14 @@ class CarTest : BehaviorSpec({
         `when`("이름이 ${Car.MAX_NAME_LENGTH}자를 초과하면") {
             then("예외가 던져진다") {
                 shouldThrow<IllegalArgumentException> {
-                    Car("a".repeat(Car.MAX_NAME_LENGTH + 1), RandomNumberGenerator())
+                    Car("a".repeat(Car.MAX_NAME_LENGTH + 1), RandomNumberGenerator)
                 }
             }
         }
         `when`("이름이 비어있으면") {
             then("예외가 던져진다") {
                 shouldThrow<IllegalArgumentException> {
-                    Car(" ", RandomNumberGenerator())
+                    Car(" ", RandomNumberGenerator)
                 }
             }
         }
@@ -32,7 +32,7 @@ class CarTest : BehaviorSpec({
             `when`("이름이 ${size}글자면") {
                 then("Car가 생성된다") {
                     shouldNotThrow<Throwable> {
-                        Car("a".repeat(size), RandomNumberGenerator())
+                        Car("a".repeat(size), RandomNumberGenerator)
                     }
                 }
             }
